+++ xxl-job-admin/src/main/java/com/xxl/job/admin/core/model/XxlJobLogGlue.java
@@ -9,7 +9,7 @@ import java.util.Date;
 public class XxlJobLogGlue {
 	
 	private int id;
-	private int jobId;				// 任务主键ID
+	private long jobId;				// 任务主键ID
 	private String glueType;		// GLUE类型	#com.xxl.job.core.glue.GlueTypeEnum
 	private String glueSource;
 	private String glueRemark;
@@ -24,11 +24,11 @@ public class XxlJobLogGlue {
 		this.id = id;
 	}
 
-	public int getJobId() {
+	public long getJobId() {
 		return jobId;
 	}
 
-	public void setJobId(int jobId) {
+	public void setJobId(long jobId) {
 		this.jobId = jobId;
 	}
